Breaking up Ising code into smaller files
The Ising.cpp file was getting too large. Separating different parts
into different cpp files.
